fix: corrected the variable type of an extern variable

+++ drivers/nxp/auth/csf_hdr_parser/csf_hdr_parser.c
@@ -39,7 +39,7 @@ static const uint8_t barker_code[CSF_BARKER_LEN] = { 0x68, 0x39, 0x27, 0x81 };
 /* Flag to indicate if values are there in rotpk_hash_table */
 bool rotpk_not_dpld =  true;
 uint8_t rotpk_hash_table[MAX_KEY_ENTRIES][SHA256_BYTES];
-uint8_t num_rotpk_hash_entries;
+uint32_t num_rotpk_hash_entries;
 
 /*
  * This function deploys the hashes of the various platform keys in
